remote-skill-scenario-4/requirements.txt
python-dotenv>=1.0.0
azure-ai-agentserver-core>=1.0.0b15
azure-ai-agentserver-langgraph>=1.0.0b15

langgraph
langchain-openai
langchain-mcp-adapters
azure-identity
typing-extensions
